Added the following features in this pull request:

- Connect patient and therapist together through the database
- Add new service interface for easy repository-pattern access
- Add criteria interface to allow more complex queries
- Add the frontend to facilitate adding, removing and view patients

+++ resources/views/admin/users/table.blade.php
@@ -1,6 +1,7 @@
 <table class="table table-hover table-outline table-striped">
     <thead class="thead-light">
         <tr>
+            <th>@lang('admin.users.table.name')</th>
             <th>@lang('admin.users.table.email')</th>
             <th>@lang('admin.users.table.actions')</th>
         </tr>
@@ -10,25 +11,38 @@
         @foreach ($users as $user)
             @if ($user->id !== Auth::id())
                 <tr>
+                    <td>{{ $user->name }}</td>
                     <td>{{ $user->email }}</td>
                     <td>
                         <a
                             class="btn btn-primary btn-sm"
-                            href="{{ url("admin/users/$user->id/edit") }}"
+                            href="{{ url(($base ?? "admin/users") . "/$user->id") }}"
                         >
-                            <i class="fas fa-edit mr-1"></i>
-                            @lang('admin.users.table.edit')
+                            <i class="fas fa-search mr-1"></i>
+                            @lang('admin.users.table.view')
                         </a>
 
-                        {!!
-                            Form::open([
-                                'class' => 'd-inline-block',
-                                'method' => 'delete',
-                                'url' => url("admin/users/$user->id"),
-                            ])
-                        !!}
-                        @include('admin.users.form-delete')
-                        {!! Form::close() !!}
+                        @can('update', $user)
+                            <a
+                                class="btn btn-primary btn-sm"
+                                href="{{ url(($base ?? "admin/users") . "/$user->id/edit") }}"
+                            >
+                                <i class="fas fa-edit mr-1"></i>
+                                @lang('admin.users.table.edit')
+                            </a>
+                        @endcan
+
+                        @can('delete', $user)
+                            {!!
+                                Form::open([
+                                    'class' => 'd-inline-block',
+                                    'method' => 'delete',
+                                    'url' => url(($base ?? "admin/users") . "/$user->id"),
+                                ])
+                            !!}
+                            @include('admin.users.form-delete')
+                            {!! Form::close() !!}
+                        @endcan
                     </td>
                 </tr>
             @endif
